% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flumisc.R
\name{epiweek_dt}
\alias{epiweek_dt}
\title{epiweek_dt: Convert an CDC week number to a calendar date}
\usage{
epiweek_dt(year, week)
}
\arguments{
\item{year}{A numeric vector}

\item{week}{A numeric vector}
}
\description{
epiweek_dt: Convert an CDC week number to a calendar date
}
