% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flu_ex.R
\docType{data}
\name{flu_ex}
\alias{flu_ex}
\title{Sample dataset for examples. Derived from CDC 122 cities dataset.}
\format{A data frame with 2847 rows and 4 variables:
\describe{
\item{year}{observation year, integer}
\item{week}{observation week, integer}
\item{fludeaths}{No. of deaths attributed to Pneumonia / Influenza}
\item{alldeaths}{No. of deaths due to any cause}
\item{perc_fludeaths}{fludeaths / alldeaths * 100}
\item{yrweek_dt}{year-week-day of rate, Date}
\item{prop_flupos}{No. of influenza positive isolates / total no. tested, from NREVSS data, see ?nrevss, the data was linked with the cdc122 data with a 2 week lead to account for delays in reporting}
}}
\source{
\url{https://data.cdc.gov/dataset/Deaths-in-122-U-S-cities-1962-2016-122-Cities-Mort/mr8w-325u}
}
\usage{
flu_ex
}
\description{
The dataset is a simple year-week dataset of influenza deaths.
}
\keyword{datasets}
