% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfunc.R
\name{flubase}
\alias{flubase}
\title{testfunc}
\usage{
flubase(dat, groups, per, pe = 0, method, indicator = "mortality", g_label)
}
\arguments{
\item{groups}{A number.}

\item{per}{a thing}

\item{pe}{another thing}

\item{method}{other thing}

\item{indicator}{of a thing}

\item{g_label}{thingy}

\item{data}{dataframe class object}
}
\value{
A table with output
}
\description{
A placeholder for functions while I build package.
}
\examples{
flubase(df, ?, ?)

}
