% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fludiff.R
\name{fludiff}
\alias{fludiff}
\title{fludiff: Compute excess mortality}
\usage{
fludiff(data = NULL, obsvar = NULL, fitvar = NULL, def = "fit",
  serfrule = F)
}
\arguments{
\item{data}{A dataframe class object, must contain observed, fitted variables}

\item{obsvar}{named observed outcome variable}

\item{fitvar}{named fitted trend variable. Can specify either fit line, or
upper threshold limit.}

\item{serfrule}{Logical, default=F. If T the function will apply the
Serfling rule which only counts consecutive observations of excess
mortality.}
}
\value{
an object of class dataframe
}
\description{
This function takes fitted data from a serflm() or similar
procedure and computes difference between observed and predicted mortality.
}
\examples{
require(flumodelr)
df <- flumodelr::fludta
serf_fit <- fluglm(fludta, outc = fludeaths , time = yrweek_dt,
offset = log(alldeaths), )  
              
excess_fludeaths <- fludiff(flu_fit, xvar=yrweek_dt, yvar=fludeaths)

}
