% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluglm.R
\name{fluglm}
\alias{fluglm}
\title{fluglm: Generate fitted prediction values with generalized linear model.}
\usage{
fluglm(data=NULL, outc=NULL, epi=NULL, time=NULL, 
              period=52, echo=F, poly=T, model.form=NULL,
              int_type="ci", alpha=0.1)
}
\arguments{
\item{data}{A dataframe class object, must contain time variable,
epidemic indicator, and measure of influenza morbidity}

\item{outc}{an unquoted name of a column in data which corresponds to
the outcome variable of interest}

\item{bl_type}{A quote string either: "season" or "viral". Denoting type of baseline
to use for fitting model.}

\item{bl_var}{an unquoted name of a variable corresponding to bl_type.}

\item{time}{an unquoted name of a column in data object, must be
a numeric/integer class variable in dataframe which corresponds
to a unit of time, must be unique (i.e. non-repeating)}

\item{period}{a numeric vector indicating period length, i.e. 52 weeks in year}

\item{echo}{A logical parameter, if T will print variables used in model.}

\item{poly}{A logical parameter, if T will include a quadratic, cubic and
quartic term.}

\item{model_form}{An object of type formula, allowing for user-specified
model to be passed on to glm(). Default missing.}

\item{int_type}{Specifies type of upper interval to be output, currently
only allows for confidence intervals. Predictional intervals to be added,
but are only approximate for Poisson families.}

\item{alpha}{The threshold for interval, default is 0.05 (one-sided).}

\item{offset}{Specify if offset term to be used, must specify log(object)}

\item{...}{options passed on to glm model (e.g. family=poisson, see ?glm)}
}
\value{
an object of class data.frame, input + y0 (fitted values), y0_ul
the upper threshold
}
\description{
Similar to serflm, but allows for more
specifications.
}
\examples{
require(flumodelr)
fludta <- flumodelr::fludta
flu_fit <- fluglm(fludta, outc = fludeaths, time = yrweek_dt)  
              
head(flu_fit)

}
\references{
Thompson WW1, Weintraub E, Dhankhar P, Cheng PY, Brammer L,
Meltzer MI, Bresee JS, Shay DK. Estimates of US influenza-associated
deaths made using four different methods. Influenza Other Respir
Viruses. 2009 Jan;3(1):37-49.
/url{https://www.ncbi.nlm.nih.gov/pubmed/19453440}
}
