% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flunb.R
\name{flunb}
\alias{flunb}
\title{flunb: Generate fitted prediction values with negative binomial
model.}
\usage{
flunb(data=NULL, outc=NULL, season, viral, time=NULL, 
              period=52, echo=F, poly=T, model_form=NULL,
              int_type="ci", alpha=0.05, offset=NULL, ...)
}
\arguments{
\item{data}{A dataframe class object, must contain time variable,
epidemic indicator, and measure of influenza morbidity}

\item{outc}{an unquoted name of a column in data which corresponds to
the outcome variable of interest}

\item{season}{Either 'T' (where the epidemic baseline model will be done based on month of year),
or the name of a column which is a logical vector flagging a given week as epidemic or not}

\item{viral}{a string vector naming 1 or more viral specimens}

\item{time}{an unquoted name of a column in data object, must be
a numeric/integer class variable in dataframe which corresponds
to a unit of time, must be unique (i.e. non-repeating)}

\item{period}{a numeric vector indicating period length, i.e. 52 weeks in year}

\item{echo}{A logical, if T will print variables used in model.}

\item{poly}{A logical, if T will include a quadratic, cubic and
quartic term.}

\item{model_form}{An object of type formula, allowing for user-specified
model to be passed on to glm(). Default missing.}

\item{int_type}{Specifies type of upper interval to be output, currently
only allows for confidence intervals. Prediction intervals to be added,
but are only approximate for Poisson families.}

\item{alpha}{The threshold for CI interval, default is 0.05 (one-sided).}

\item{offset}{Specify if offset term to be used, must specify log(object)}

\item{...}{other options passed on to glm model (e.g. family=poisson, see ?glm)}
}
\value{
an object of class data.frame, fit, upper and lower confidence bounds
}
\description{
Similar to fluglm, but glm does not have a negative
binomial option so this function calls MASS::glm.nb separately.
}
\examples{
require(flumodelr)
fludta <- flumodelr::fludta
flu_fit <- flunb(fludta, outc = fludeaths, time = yrweek_dt)  
              
head(flu_fit)

}
\references{
Thompson WW1, Weintraub E, Dhankhar P, Cheng PY, Brammer L,
Meltzer MI, Bresee JS, Shay DK. Estimates of US influenza-associated
deaths made using four different methods. Influenza Other Respir
Viruses. 2009 Jan;3(1):37-49.
/url{https://www.ncbi.nlm.nih.gov/pubmed/19453440}
}
