% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flunnet.R
\name{flunnet}
\alias{flunnet}
\title{flunnet: Estimating attributable mortality with Neural Networks}
\usage{
flunnet()
}
\arguments{
\item{data}{A dataframe class object, must contain time variable,
epidemic indicator, and measure of influenza morbidity}

\item{outc}{an unquoted name of a column in data which corresponds to
the outcome variable of interest}

\item{time}{an unquoted name of a column in data object, must be
a numeric/integer class variable in dataframe which corresponds
to a unit of time, must be unique (i.e. non-repeating)}

\item{season}{Either 'T', where the epidemic baseline model will be created
assuming time variable is a date, and Oct-May is epidemic season, or the
name of a column which is a logical vector flagging a given week as
epidemic or not}

\item{echo}{A logical, if T will print variables used in model.}

\item{...}{other options passed on to \code{nnetar()}}
}
\value{
an object of class data.frame, fit, upper and lower confidence bounds
}
\description{
Estimates fitted curves and baseline using the \code{nnetar} function
in the forecast package. This is a Feed-forward neural network
with a single hidden layer and lagged inputs for forecasting
univariate time series.
}
\examples{

}
