% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluplot.R
\name{fluplot}
\alias{fluplot}
\title{fluplot: Convenience wrapper for standard trend plot}
\usage{
fluplot(data = NULL, xvar = NULL, yvar = NULL, baseline = y0,
  threshold = y0_ul, linenames = NULL, ylab = "[OUTCOME]",
  title = "[INSERT TITLE]")
}
\arguments{
\item{data}{A dataframe class object, must contain time variable,
outcome variable}

\item{xvar}{x-axis variable, must be date class}

\item{yvar}{y-axis, outcome variable}

\item{baseline}{second line variable, fitted cyclical trend. Default="y0".}

\item{threshold}{third line, upper threshold for cyclical trend. Default="y0_ul".}

\item{linenames}{Defaults specified, but can override. Must be character
vector of length=3.}

\item{ylab}{Character vector for Y-axis label}

\item{title}{Character vector for Plot Title}
}
\value{
an object of class gg, ggplot.
}
\description{
This models wraps ggplot2 commands into a standard
form for making cyclical trend line plots of influenza. Developed
only for convenience, the authors encourage individuals to make thier
own customized graphs as appropriate.
}
\examples{
require(flumodelr)
flu_ex <- flumodelr::flu_ex
flu_fit <- serflm(flu_ex, outc = "fludeaths", time = "yrweek_dt")  
              
fluplot(flu_fit, x="yrweek_dt", y="fludeaths")

}
