% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fitvals.R
\name{gam_fitvals}
\alias{gam_fitvals}
\title{gam_fitvals: extract fit values from glm fit object}
\usage{
gam_fitvals(gam.fit, data, alpha=0.05)
}
\arguments{
\item{alpha}{The threshold for CI interval, default is 0.05 (one-sided).}
}
\description{
Helper function which extracts needed parameters from gam
fit object
}
