% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fitvals.R
\name{glm_fitvals}
\alias{glm_fitvals}
\title{glm_fitvals: extract fit values from glm fit object}
\usage{
glm_fitvals(glm.fit, data, alpha=0.05)
}
\arguments{
\item{alpha}{The threshold for CI interval, default is 0.05 (one-sided).}

\item{offset}{Specify if offset term to be used, must specify log(object)}
}
\description{
Helper function which extracts needed parameters from glm
fit object
}
