% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ird.R
\name{gr}
\alias{gr}
\title{gr: Line graph for ird models}
\usage{
gr(data)
}
\arguments{
\item{data}{A formatted dataset created irb() rb() commands}
}
\description{
Outputs simple line graph for incidence rate difference
models
}
