% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ird.R
\name{ird}
\alias{ird}
\title{ird: Perform incidence rate-difference models}
\usage{
ird(data = NULL, flu = NULL, epi = NULL, time = NULL,
  t.interval = "wofy")
}
\arguments{
\item{data}{A dataframe class object, must contain time variable,
epidemic indicator, and measure of influenza morbidity}

\item{flu}{A character string of length=1, identifies
a numeric type variable in dataframe data which is the measure
of disease morbidity / mortality}

\item{epi}{A character string of length=1, identifies
a logical type variable in dataframe data which indicates epidemic
vs non-epidemic periods}

\item{time}{A character string of length=1, identifies
a numeric/integer class variable in dataframe which corresponds
to a unit of time, must be unique (i.e. non-repeating)}

\item{t.interval}{A character string, must specify whether
unit of time cycle is weeks/years ("wofy", i.e. 52 weeks),
or month/years ("mofy"), days/years("dofy")}

\item{t.respStart}{A character string, must specify whether
unit of time cycle is weeks/years ("wofy", i.e. 52 weeks),
or month/years ("mofy"), days/years("dofy")}

\item{t.respStop}{A character string, must specify whether
unit of time cycle is weeks/years ("wofy", i.e. 52 weeks),
or month/years ("mofy"), days/years("dofy")}
}
\value{
dataframe
}
\description{
Calculates the rate difference between influenza season
and peri-influenza season, or between influenza season and
summer season
}
\examples{
require(flumodelr)
flu_ex <- flumodelr::flu_ex
fit <- serflm(data=flu_ex, 
              flu = "fludeaths", epi="epi", 
              time="yrweek_dt")
              
summary(fit)

}
\references{
Serfling RE. Methods for current statistical analysis of
excess pneumonia-influenza deaths. Public Health Rep. 1963 Jun;
78(6): 494 - 506.
/url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC1915276/}
}
