% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_flu_form.R
\name{mk_flu_form}
\alias{mk_flu_form}
\title{mk_flu_form: Build model formula for internal use}
\usage{
mk_flu_form(outc, offset, poly, model_form)
}
\value{
an object of class formula
}
\description{
This function is not meant to be called directly, it takes input
from user-called functions and builds a formula class object for use in
regression commands.
}
