% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serflm.R
\name{serflm}
\alias{serflm}
\title{serflm: Fit a Serling Model on Time Series Data}
\usage{
serflm(data = NULL, outc = NULL, epi = NULL, time = NULL,
  t.interval = "wofy")
}
\arguments{
\item{data}{A dataframe class object, must contain time variable,
epidemic indicator, and measure of influenza morbidity}

\item{outc}{A character string of length=1, identifies
a numeric type variable in dataframe data which is the measure
of disease morbidity / mortality}

\item{epi}{either a character string of length=1 (e.g. '"epi")',
identifying a logical type variable in dataframe data which indicates epidemic
vs non-epidemic periods, or if unspecified will default to Sept - May.}

\item{time}{A character string of length=1, identifies
a numeric/integer class variable in dataframe which corresponds
to a unit of time, must be unique (i.e. non-repeating)}

\item{t.interval}{A character string, must specify whether
unit of time cycle is weeks/years ("wofy", i.e. 52 weeks),
or month/years ("mofy"), days/years("dofy")}
}
\value{
an object of class data.frame, input + y0 (fitted values), y0_ul
the upper serfling threshold
}
\description{
Performs a cyclical linear regression model,
or 'serfling' model
}
\examples{
require(flumodelr)
flu_ex <- flumodelr::flu_ex
fit <- serflm(data=flu_ex, outc = "fludeaths", time="yrweek_dt")
              
summary(fit)

}
\references{
Serfling RE. Methods for current statistical analysis of
excess pneumonia-influenza deaths. Public Health Rep. 1963 Jun;
78(6): 494 - 506.
/url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC1915276/}
}
