% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flubase.R
\name{flubase}
\alias{flubase}
\title{flubase: An internal function for taking the average of previous
baseline seasons}
\usage{
flubase(data, outc, time, season)
}
\arguments{
\item{data}{A dataset which contains a influenza measure, time and grouping
variable, season}

\item{outc}{Influenza measure}

\item{time}{A a numeric vector with influenza measure}

\item{group}{An vector ordered by time which indicates grouping by season '}

\item{k}{The order of lookback, if k=5 then will take mean of previous 5
groups (i.e. influenza seasons)}
}
\value{
a vector of length equal to x
}
\description{
A simple function will compute the mean by group, across a time unit
}
