% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluglm.R
\name{fluglm}
\alias{fluglm}
\title{fluglm: Generate fitted prediction values with generalized linear model.}
\usage{
fluglm(data=NULL, outc=NULL, season, viral, 
              time=NULL, period=52, echo=F, poly=T, 
              model_form='none', int_type="ci", alpha=0.05, offset,
              glmnb=F, ...)
}
\arguments{
\item{data}{A dataframe class object, must contain time variable,
epidemic indicator, and measure of influenza morbidity}

\item{outc}{an unquoted name of a column in data which corresponds to
the outcome variable of interest}

\item{season}{Either 'T', where the epidemic baseline model will be created
assuming time variable is a date, and Oct-May is epidemic season, or the
name of a column which is a logical vector flagging a given week as
epidemic or not}

\item{viral}{a string vector naming 1 or more viral specimens}

\item{time}{an unquoted name of a column in data object, must be
a numeric/integer class variable in dataframe which corresponds
to a unit of time, must be unique (i.e. non-repeating)}

\item{period}{a numeric vector indicating period length, i.e. 52 weeks in year}

\item{echo}{A logical, if T will print variables used in model.}

\item{poly}{A logical, if T will include a quadratic, cubic,
quartic and quintic term.}

\item{model_form}{An object of type formula, allowing for user-specified
model to be passed on to glm(). Default missing.}

\item{int_type}{Specifies type of upper interval to be output, currently
only allows for confidence intervals. Prediction intervals to be added,
but are only approximate for Poisson families.}

\item{alpha}{The threshold for CI interval, default is 0.05 (one-sided).}

\item{offset}{Specify if offset term to be used, must specify log(object)}

\item{glmnb}{Logical, if True will run a negative binomial model using
MASS::nb.glm}

\item{...}{other options passed on to glm model (e.g. family=poisson, see ?glm)}
}
\value{
an object of class data.frame, fit, upper and lower confidence bounds
}
\description{
Similar to serflm, but allows for more
specifications.
}
\examples{
require(flumodelr)
fludta <- flumodelr::fludta
flu_fit <- fluglm(fludta, outc = fludeaths, time = week_in_order, season=epi)  
              
head(flu_fit)

}
\references{
Thompson WW1, Weintraub E, Dhankhar P, Cheng PY, Brammer L,
Meltzer MI, Bresee JS, Shay DK. Estimates of US influenza-associated
deaths made using four different methods. Influenza Other Respir
Viruses. 2009 Jan;3(1):37-49.
/url{https://www.ncbi.nlm.nih.gov/pubmed/19453440}
}
