% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flum.R
\name{flum}
\alias{flum}
\title{flum: General modeling function for Attributable Influenza Mortality}
\usage{
flum(data=NULL, model="serflm", ...)
}
\arguments{
\item{data}{A dataframe class object, must contain time variable,
epidemic indicator, and measure of influenza morbidity}

\item{model}{A character of length 1, indicates which model to fit. Options
include "ird" (see ?ird), "serflm" (see ?serflm), virology model (vignette('05-modserf',
package = "flumodelr")), "arima" (see ?flurima).}

\item{...}{Extra arguments depending on model requested, see examples and vignettes.}
}
\value{
an object of class data.frame, input + y0 (fitted values), threshold
if applicable, and computed excess morbidity
}
\description{
General function wrapper to perform a variety of forecasted
models. The output will always be a set of fitted values.
}
\examples{
require(flumodelr)
fludta <- flumodelr::fludta
flu_fit <- flum(fludta, model="serflm", outc = fludeaths, time = yrweek_dt)  
              
head(flu_fit)

}
\references{
Thompson WW1, Weintraub E, Dhankhar P, Cheng PY, Brammer L,
Meltzer MI, Bresee JS, Shay DK. Estimates of US influenza-associated
deaths made using four different methods.
Influenza Other Respir Viruses. 2009 Jan;3(1):37-49.
/url{https://www.ncbi.nlm.nih.gov/pubmed/19453440}
}
