% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flumodelr.R
\docType{package}
\name{flumodelr}
\alias{flumodelr}
\alias{flumodelr-package}
\title{flumodelr: 
An R Package for Estimating Attributable Influenza Morbidity and Mortality.}
\description{
This package contains multiple vignettes describing datasources, 
background and examples.
}
\section{Vignettes}{

Introduction: A general introduction to package 
testdata: A description of example data sources
}

