% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serflm.R
\name{serflm}
\alias{serflm}
\title{serflm}
\usage{
serflm(data = NULL)
}
\arguments{
\item{data}{A dataframe class object}
}
\value{
an object of class lm.summary
}
\description{
Performs a cyclical linear regression model,
or 'serfling' model
}
\examples{
require(flumodelr)
cdc122 <- flumodelr::cdc122city
fit <- serflm(data=cdc122)
summary(fit)

}
\references{
Serfling RE. Methods for current statistical analysis of
excess pneumonia-influenza deaths. Public Health Rep. 1963 Jun;
78(6): 494 - 506.
/url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC1915276/}
}
