% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serflm.R
\name{serflm}
\alias{serflm}
\title{serflm}
\usage{
serflm(data = NULL)
}
\arguments{
\item{data}{A dataframe class object}
}
\value{
an object of class lm.summary
}
\description{
Performs a cyclical linear regression model, 
or 'serfling' model
}
\examples{
require(flumodelr)
cdc122 <- flumodelr::cdc122city
m.out <- serflm(treat ~ age + educ + black + hispan, 
                data=cdc122)
summary(m.out)


}
